Remove Api project DotNet9 makes all the packages available already Merge functionalities Increment Major version
