+++ ActionResponse.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActionResponse", "ActionResponse\ActionResponse.csproj", "{79A549B2-EDCA-4E6E-9E56-08356A5CEEB1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ActionResponse_Api", "ActionResponse_Api\ActionResponse_Api.csproj", "{6E3F6CF6-6250-485C-BAE2-222D0B8F2A80}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{79A549B2-EDCA-4E6E-9E56-08356A5CEEB1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{79A549B2-EDCA-4E6E-9E56-08356A5CEEB1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{79A549B2-EDCA-4E6E-9E56-08356A5CEEB1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6E3F6CF6-6250-485C-BAE2-222D0B8F2A80}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6E3F6CF6-6250-485C-BAE2-222D0B8F2A80}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6E3F6CF6-6250-485C-BAE2-222D0B8F2A80}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6E3F6CF6-6250-485C-BAE2-222D0B8F2A80}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
