+++ ActionResponse/ActionResponse.csproj
@@ -13,11 +13,12 @@
         <PackageTags>C#, Response, API</PackageTags>
         <NeutralLanguage>en</NeutralLanguage>
         <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <Version>2.0.1</Version>
+        <Version>3.0.0</Version>
         <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
         <IsPackable>true</IsPackable>
         <ProduceReferenceAssembly>False</ProduceReferenceAssembly>
         <PackageOutputPath>bin\Nuget</PackageOutputPath>
+        <NuGetAudit>False</NuGetAudit>
     </PropertyGroup>
 
     <ItemGroup>
