+++ ActionResponse/Properties/launchSettings.json
@@ -1,0 +1,12 @@
+{
+  "profiles": {
+    "ActionResponse": {
+      "commandName": "Project",
+      "launchBrowser": true,
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      },
+      "applicationUrl": "https://localhost:60150;http://localhost:60151"
+    }
+  }
+}
