--- ActionResponse_Api/ActionResponse_Api.csproj
@@ -1,34 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net9.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <OutputType>Library</OutputType>
-        <Authors>Cubelaster</Authors>
-        <Description>A package meant to act as a wrapper for API calls.</Description>
-        <PackageProjectUrl>https://github.com/Cubelaster/ActionResponse</PackageProjectUrl>
-        <PackageReadmeFile>README.md</PackageReadmeFile>
-        <RepositoryUrl>https://github.com/Cubelaster/ActionResponse</RepositoryUrl>
-        <PackageTags>C#, Response, API</PackageTags>
-        <NeutralLanguage>en</NeutralLanguage>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-        <IsPackable>true</IsPackable>
-        <ProduceReferenceAssembly>False</ProduceReferenceAssembly>
-        <PackageOutputPath>bin\Nuget</PackageOutputPath>
-        <Version>2.0.1</Version>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <None Include="..\README.md">
-            <Pack>True</Pack>
-            <PackagePath></PackagePath>
-        </None>
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\ActionResponse\ActionResponse.csproj" />
-    </ItemGroup>
-
-</Project>
